Implement Landing page with functional park randomization
