+++ src/main/resources/templates/landing.html
@@ -21,6 +21,27 @@
 
 <!--TODO: LANDING CONTENT GOES HERE -->
 
+
+<div th:each="park : ${shuffledNationalParks}">
+    <div>
+        <a th:href="@{'/parks/parkcard/' + ${park.parkCode}}" th:text="${park.fullName}"></a>
+        <h1 th:text="${park.fullName}">Default Name</h1>
+        <div th:each="image : ${park.images[0]}">
+            <img th:src="${image.url}" width="400" height="200"/>
+        </div>
+    </div>
+</div>
+
+
+<!--<div th:each="park : ${npsResponse.data}">-->
+<!--    <div th:if = "${park.designation.equals('National Park')}">-->
+<!--        <a th:href="@{'/parks/parkcard/' + ${park.parkCode}}" th:text="${park.fullName}"></a>-->
+<!--        <h1 th:text="${park.fullName}">Default Name</h1>-->
+<!--        <div th:each=" image : ${park.images[0]}">-->
+<!--            <img th:src="${image.url}" width="400" height="200"/>-->
+<!--        </div>-->
+
+
 </body>
 </html>
